move out lambda operator code to separate file
