+++ bentoml/yatai/deployment/operator.py
@@ -28,7 +28,9 @@ def get_deployment_operator(yatai_service, deployment_pb):
 
         return SageMakerDeploymentOperator(yatai_service)
     elif operator == DeploymentSpec.AWS_LAMBDA:
-        from bentoml.yatai.deployment.aws_lambda import AwsLambdaDeploymentOperator
+        from bentoml.yatai.deployment.aws_lambda.operator import (
+            AwsLambdaDeploymentOperator,
+        )
 
         return AwsLambdaDeploymentOperator(yatai_service)
     elif operator == DeploymentSpec.GCP_FUNCTION:
